requirements.txt
pandas>=2.0.0
notion-client>=2.2.1
dropbox>=11.36.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0
openpyxl>=3.1.0
schedule>=1.2.0